+++ 00_global/01_libft/ft_memset.c
@@ -1,11 +1,9 @@
-#include <stdio.h>
-#include <stdlib.h>
 #include "libft.h"
 
-void* ft_memset(void *b, int c, unsigned int len)
+void* ft_memset(void *b, int c, size_t len)
 {
     unsigned char* char_b = (unsigned char*) b;
-    unsigned int i = 0;
+    size_t i = 0;
     while (i < len)
     {
         char_b[i] = (unsigned char)c;
